map refactored to typescript and organized for easy access. dotenv plugged. map size set in css. added google maps react wrapper and types in the package.json.
