+++ client/app/AppRoutes.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { Route, Routes } from "react-router-dom";
 import Home from "../features/home/Home";
+import MapApp from "../features/mapApp/mapComponent/mapApp";
 
 /**
  * COMPONENT
@@ -10,7 +11,8 @@ const AppRoutes = () => {
   return (
     <div>
       <Routes>
-        <Route path="/*" element={<Home />} />
+        <Route path="*" element={<Home />} />
+        <Route path="/map" element={<MapApp />} />
       </Routes>
     </div>
   );
