+++ client/features/navbar/Navbar.tsx
@@ -1,4 +1,5 @@
 import * as React from "react";
+import { NavLink } from "react-router-dom";
 
 const Navbar = () => {
   return (
@@ -6,7 +7,12 @@ const Navbar = () => {
       <h1>Equidist</h1>
       <nav>
         <h3>LOGO</h3>
-        <button>NAV BUTTON</button>
+        <button>
+          <NavLink to="/map">Click Here For Map</NavLink>
+        </button>
+        <button>
+          <NavLink to="/home">Click Here for Home</NavLink>
+        </button>
       </nav>
       <hr />
     </div>
